Only run mitmproxy setup if not done already.

+++ bin/setup.sh
@@ -86,11 +86,13 @@ sed -i -e "s/ip = 192.168.56.1/ip = $HOSTIP/" $ROOTDIR/cuckoo.conf
 
 sudo cp ~/cuckoo-tools/files/suricata.yaml /etc/suricata
 
-echo -n "Fix mitmproxy. "
-mitmproxy >> ~/src/cuckoo/log/mitmproxy 2>&1 &
-MITM_PID=$!
-sleep 10
-kill -9 $MITM_PID > /dev/null 2>&1 | true
-cp ~/.mitmproxy/mitmproxy-ca-cert.p12 ~/src/cuckoo/analyzer/windows/bin/cert.p12
-echo "Done."
+if [ ! -e ~/src/cuckoo/analyzer/windows/bin/cert.p12 ]; then
+    echo -n "Fix mitmproxy. "
+    mitmproxy >> ~/src/cuckoo/log/mitmproxy 2>&1 &
+    MITM_PID=$!
+    sleep 10
+    kill -9 $MITM_PID > /dev/null 2>&1 | true
+    cp ~/.mitmproxy/mitmproxy-ca-cert.p12 ~/src/cuckoo/analyzer/windows/bin/cert.p12
+    echo "Done."
+fi
